pre-build the SMF repository in the disk image
I have been trying to avoid this, because doing the import live via EMI as a
system would has found many bugs, but it is inarguably slow as molasses under
emulation.

(This unfortunately has the side effect that we don't get EMI output
indicating that something is happening during boot while we _do_ still have a
delay.  Sigh)