install _everything_ into the disk image (FLAG DAY)
This installs everything we have, via the `osnet-redistributable` package into
the disk image.  This radically increases the size of
our boot archive (especially with DWARF) and requires a current illumos
arm64-gate to boot.

With this everything that illumos builds and puts in packages goes into the
disk image.  It is important to note that some (perhaps large) quantity of
this doesn't work, is missing dependencies (because we have no `pkgdepend(1)`
as yet), or is otherwise broken.