sysroot: Add packages necessary for dependency resolution [FLAG DAY]
Because of a bug in on the illumos side that was working, but only in
limited cases for the wrong reasons, the sysroot was missing a couple of
packages that are _actually_ necessary to resolve dependencies.

Add them to the Makefile, and fully-specify packages while in there.

You will need to run `make sysroot` again after you've pulled this
change, and before building the matching illumos change.