Support EFI disk images
EFI disk images are easier to create and are also allow
easy expansion of the root ZFS pool to fill the available
disk.

This adds a new set of flags to the disk image script and
either --mbr or --efi must be specified to select the format
of the disk image which will be produced.